requirements.txt
PyMuPDF==1.23.17
gradio==4.11.0      
langchain==0.2.1
langchain-weaviate==0.0.2
langchain-community==0.2.1
langchain-openai==0.1.7
Pillow==10.1.0      
torch==2.1.1        
transformers==4.35.2
PyYAML==6.0.1
weaviate-client==4.6.3
pypdf==4.0.0
Jinja2==3.1.3
accelerate==0.26.1
sentence-transformers==2.2.2
pydantic-settings==2.2.1
tiktoken==0.7.0
